+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - widgets
     - script: bypass keyboard shortcuts already used by the editor if it is enabled
     - matrix: removed ambiguous `start` property
+    - matrix: advanced syntax blocks can be passed to children without being resolved at the matrix\' scope by adding an underscore before the opening bracket
     - text: add `soft` mode for the `wrap` property
 
 ## 1.9.6